# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Aplicaciones Web II",
        "description": "Documentacion generada automaticamente con L5-Swagger",
        "version": "1.0"
    },
    "paths": {
        "/api/estudiantes": {
            "get": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Listar estudiantes",
                "operationId": "209e318b7ac388b93bccf7af4162e33b",
                "responses": {
                    "200": {
                        "description": "Lista de Estudiantes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Crear un nuevo estudiante",
                "operationId": "bf7df90391b947c8dc2f77570e61358a",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Indica que se espera una respuesta en formato JSON",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "cedula",
                                    "correo",
                                    "paralelo_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Perez"
                                    },
                                    "cedula": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "paralelo_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estudiante creado exitosamente"
                    },
                    "422": {
                        "description": "Errores de validacion"
                    }
                }
            }
        },
        "/api/estudiantes/{id}": {
            "put": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Actualizar un estudiante existente",
                "operationId": "a0d23067013987f1a070c064758e522b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del estudiante a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Pedro Gómez"
                                    },
                                    "cedula": {
                                        "type": "string",
                                        "example": "1101234568"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "pedro@example.com"
                                    },
                                    "paralelo_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estudiante actualizado correctamente"
                    },
                    "404": {
                        "description": "Estudiante no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Eliminar un estudiante",
                "operationId": "caeee49dd1fc2f0619a1069d2e414f0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del estudiante a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estudiante eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Estudiante eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estudiante no encontrado"
                    }
                }
            }
        },
        "/api/paralelos": {
            "get": {
                "tags": [
                    "Paralelos"
                ],
                "summary": "Listar todos los paralelos",
                "operationId": "30409aa4e072067e1395716421b89f6c",
                "responses": {
                    "200": {
                        "description": "Listado de paralelos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Paralelo"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Paralelos"
                ],
                "summary": "Crear un nuevo paralelo",
                "operationId": "63f2e6de83fe35723b5077a59ec0f619",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Paralelo A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paralelo creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Paralelo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/paralelos/{id}": {
            "get": {
                "tags": [
                    "Paralelos"
                ],
                "summary": "Mostrar un paralelo por ID",
                "operationId": "4d4a75bdf60cae9e6845622d9d5089d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paralelo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paralelo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Paralelo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paralelo no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Paralelos"
                ],
                "summary": "Actualizar un paralelo",
                "operationId": "09c0363579e6f12d0cf5d6a8934f4fad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paralelo a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Paralelo B"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paralelo actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Paralelo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paralelo no encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Paralelos"
                ],
                "summary": "Eliminar un paralelo",
                "operationId": "e1db2fb826fbe8f0eb7a8c73c9222f49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del paralelo a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paralelo eliminado exitosamente"
                    },
                    "404": {
                        "description": "Paralelo no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Paralelo": {
                "title": "Paralelo",
                "required": [
                    "id",
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Paralelo A"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Estudiantes",
            "description": "Estudiantes"
        },
        {
            "name": "Paralelos",
            "description": "Paralelos"
        }
    ]
}